+++ nringthymeleaf/src/main/java/com/katamlek/nringthymeleaf/domain/UserSalesTarget.java
@@ -5,6 +5,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Calendar;
 
 @Entity
@@ -19,5 +20,5 @@ public class UserSalesTarget {
 
     @Temporal(TemporalType.DATE) // Entered automatically as the last day of the month
     private Calendar periodTo;
-    private Long targetAmount;
+    private BigDecimal targetAmount;
 }
